Move rollbar reporter into a separate module
