+++ lib/pliny/error_reporter.rb
@@ -1,5 +1,4 @@
-require 'rollbar/exception_reporter'
-require 'rollbar/request_data_extractor'
+require_relative 'error_reporters/rollbar'
 
 class Pliny::ErrorReporter
   def self.notify(exception, context: {}, rack_env: {})
@@ -14,29 +13,5 @@ class Pliny::ErrorReporter
     end
   end
 
-  class RollbarReporter
-    include ::Rollbar::ExceptionReporter
-    include ::Rollbar::RequestDataExtractor
-
-    def notify(exception, context:, rack_env:)
-      Rollbar.reset_notifier!
-      scope = fetch_scope(context: context, rack_env: rack_env)
-      Rollbar.scoped(scope) do
-        report_exception_to_rollbar(rack_env, exception)
-      end
-    end
-
-    private
-
-    def fetch_scope(context:, rack_env:)
-      {
-        request: proc { extract_request_data_from_rack(rack_env) }
-      }
-    rescue Exception => e
-      report_exception_to_rollbar(rack_env, e)
-      raise
-    end
-  end
-
-  REPORTERS = [RollbarReporter]
+  REPORTERS = [Pliny::ErrorReporters::Rollbar]
 end
