+++ lib/template/config/initializers/rollbar.rb
@@ -1,3 +1,5 @@
+require 'pliny/error_reporters/rollbar'
+
 Rollbar.configure do |config|
   config.enabled = ENV.has_key?('ROLLBAR_ACCESS_TOKEN')
   config.disable_rack_monkey_patch = true
